diagnose-error schema to shared/tools
